signal: make windows docs show up on unix builds
